refactor: move yajl_buf to buffer
