+++ crates/yajl/src/parser.rs
@@ -4,8 +4,8 @@ pub(crate) use parser_impl::yajl_parse_integer;
 use parser_impl::{ByteStack, ParseState};
 
 use crate::{
+    buffer::Buffer,
     yajl_alloc::{yajl_alloc_funcs, yajl_set_default_alloc_funcs},
-    yajl_buf::Buffer,
     yajl_lex::{yajl_lex_alloc, yajl_lex_free, yajl_lexer_t},
     Status,
 };
